requirements.txt
fastapi[all]==0.92.0
fastapi-sqla==2.8.1
fastapi-users[all]==10.4.0
uvicorn==0.20.0
asyncpg==0.27.0
psycopg2-binary==2.9.5
pydantic==1.10.5
SQLAlchemy==1.4.46
aiohttp==3.8.4
aiofiles==23.1.0
aioredis==2.0.1
alembic==1.9.4
pytest==7.2.1
jmespath==1.0.1
PyYAML==6.0
lxml==4.9.2
syncer==1.3.0
dramatiq[rabbitmq, watch]==1.14.1
python-multipart==0.0.5
plumbum==1.8.1
tap.py==3.1
aiohttp_retry==2.8.3
jinja2==3.1.2
pgpy==0.6.0
markdown==3.4.1
fastapi-users-db-sqlalchemy==4.0.5
httpx-oauth==0.11.0
sentry-sdk[fastapi]==1.15.0  # requires FastAPI>=0.79.0
git+https://git.almalinux.org/almalinux/cas_wrapper.git@0.0.6#egg=cas_wrapper
